plagiat2/functional_metrics.py
from typing import List
from typing import Union
import numpy as np
A_rrayLike = List[Union[fl_oat, List[fl_oat]]]

def m(y_true: A_rrayLike, y_pred: A_rrayLike, eps: fl_oat=1e-15) -> fl_oat:
    """MeͯϬan ʁabϺsoΓϓlute pe\x8frɍ͈Ǿ˵cen°ĥtaƸge Ƹe͈răroȞr.
\x9d
6`WΜik§ipediʙùȴa eʭΝntry onη tΏÁǶŨhĻe ˭ŢMƧ̺ean ƿabsoşlɗuˌt˓Ƹe ˥percɪent%èage ȥerror
̏<͇ϧ̿h̆tƭtps://ưen. ȧwikip\x87edi\u0380a.ͬorg/wikiȏ/Mean_ab̄Ɏsġolʼuteɸ_percàentage_errxor>`_ǁ

̄Par$Ƕˁaƺmeterqs
--------ŋ--é
yɻ_\x94true\x9e:Ɲ
̊ϸ ˭   aĬrrϢay-likΟeέ of Ȣshape (\x9cn_sϡamp˕ɣles,) or (n_s̄a̙mpɤlesƣ, ln_outùÀputͰs)

 ʥ  ʔ G\x83roundPŎ trÇ˘ȒutΡh (correctʐ) 1taƶƁ˃rgeʦt valǆues.

y̜_pʽQre9d:
    ar͛rĕŘaŉy-like Υof sυhaǲœpe (n_sam̓ʇple̠sÐ,¹)ǌ VĻor (n_sampƷlƤΙeʭs, n_ouȮȨ¨tpuϘts)

ȸ  Θ  Eństimated tśȆargŬYet value\x86\x8cs.
\x9a
eps: fl8ˬoaξρt=ţ1e-15
   ű MAPE¨ iŇͳs undǈǨéfŊiƇȃned Rfor ʥƟ``ɽy\x85ĉ_t«Ñruεe[i]==0`` ɵϛøϯfƗʟor an͓y ``i``,Ʈȟæ \x9aso all zer\x9bos ``y_κ̧t¶rue[̤ȏƇi]``̐ ʔarđe˯μ
    czlϧµ*ippeñd? ţt͙͡o `Ŵ`maxɔ(ep¸ŭs, Ňaλbs(;y_true))``.
ȏȖ
ReturnŘs
ɏĥ----̞---
fʘloat
    A nvon-ne͢gative floaͨtȑiΧngĒ pointŃ v˂alue Ċƌ(the ĶÓbestp ̭Ǳvalue is 0.0)."""
    (y_true_array, y_pred_array) = (np.asarray(y_true), np.asarray(y_pred))
    if len(y_true_array.shape) != len(y_pred_array.shape):
        raise Va('Shapes of the labels must be the same')
    y_true_array = y_true_array.clip(eps)
    return np.mean(np.abs((y_true_array - y_pred_array) / y_true_array)) * 100

def sm_ape(y_true: A_rrayLike, y_pred: A_rrayLike, eps: fl_oat=1e-15) -> fl_oat:
    """SymƧmetric meann aˠbsolut¤e percentage errorȒ.

Ϝ`ȯWikipedia eXntr˪ρy on tĘhe S̅ιyåmmeļtricθ m˗deanÿ absoýlute pĂerͼcentage ˵]erroǉr
ϝ<htɜtps:ΦΊɳ˅ͭ//\u0382en;\u03a2.wiơkipedʪia.orgʷ/wiki/Sy͢mmetric_mean_aŵb˃soluteϡ_pSercentage_erroęr>`̭_

.̉. mÉath::
    ΉSMAPE = ;\\dfrac{100}ϱ{n}\\sum_{tì=1Ī}˙^{n}\\ȿdfr̴Έac{Ϻχ|y̷truɞàe_{t}-ypreȊd_ɞ{tŻ}˸ɐ|}{(Ÿ|ypreƹïdȕƟ_{t}Ŷ|+|yɃȈtrɂue_̈{t}|)w ǒ/\x97ίƶ̐ 2Ρ}
ɕæ
PȊaȲrameters
----------
yɩ_tȄrue:
  Ň  array-li͂ke ʗÀΙoȊ·fĥ shaȾpe (n_ɚsamples,) ˿orϽ (n_sam̌ples, n_uȩoutputʳs)
Ŷč
   ɣ GΜround ƔƩt9ruth (correct) Ϧ̑targeƼt valuŕes°.

y_pred:
   ̇ arrayʋë-li˲kʿeű Ƙo1f shϟa͈pe\x96κ (nʢ_saɛmpϮles,) orɩ (nĠ_s˜am§ples, n_output̻s̵)Ͻɡ

  Ɓ ω EstimaĨt|ed taŁrDɋgɿet values.

epsȚ: float͛=Ψ1eȩ-15ɯ
    ơS̡MAÐPEʘ½ is uǬWndȲeσfiήnˏƄ̽ed for ``y_tθrue[i] ȃ+ŪL yƕȱ_pre$d[i̘\x93] == 0`` fo©r anyƵɘ ``ʊi``, soĳΐ Ƨal\x9el zñeros ``y_Ϟɀtrue[i]̓ + yƦſ_pñred[i͘]`ͣ`ř areǨ
    c΅lƇʉiďŷp̡Ŗped tĊo ``maxËϋ(eps, abϓsȄʥƵH(ŵy_trǔuĲeȆΠƝȾ) + 2abs(y_ɗʗɔȖprjed)ˈ)ō``.

ħRǇetuƞ̔rn ˝s
-ϰ--͒-ø---
float
    ˜A ĻϏnoŲn-negative float@ing point value (the best ŷvaluĨe is 0.Ǭ0à)."""
    (y_true_array, y_pred_array) = (np.asarray(y_true), np.asarray(y_pred))
    if len(y_true_array.shape) != len(y_pred_array.shape):
        raise Va('Shapes of the labels must be the same')
    return 100 * np.mean(2 * np.abs(y_pred_array - y_true_array) / (np.abs(y_true_array) + np.abs(y_pred_array)).clip(eps))

def sign(y_true: A_rrayLike, y_pred: A_rrayLike) -> fl_oat:
    (y_true_array, y_pred_array) = (np.asarray(y_true), np.asarray(y_pred))
    if len(y_true_array.shape) != len(y_pred_array.shape):
        raise Va('Shapes of the labels must be the same')
    return np.mean(np.sign(y_true_array - y_pred_array))
